python_courses/alura_djangoREST/requirements.txt
asgiref==3.8.1
Django==5.0.3
django-filter==24.3
djangorestframework==3.15.2
Faker==33.1.0
Markdown==3.7
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
validate-docbr==1.10.0
